refactor: move ActionMethods into ParcelChangeMethods
There used to be more action methods so a separate file was clearer.
Now theres only one, so move it with in with its peers
