+++ packages/dataparcels/src/parcel/__test__/ParcelChangeMethods-test.js
@@ -4,6 +4,31 @@ import GetAction from '../../util/__test__/GetAction-testUtil';
 import ParcelShape from '../../parcelShape/ParcelShape';
 import TestValidateValueUpdater from '../../util/__test__/TestValidateValueUpdater-testUtil';
 
+test('Parcel.dispatch() should pass handleChange to newly created parcel', () => {
+    let handleChange = jest.fn();
+
+    let parcel = new Parcel({
+        value: 123,
+        handleChange
+    });
+
+    parcel.onChange(456);
+
+    let [newParcel, changeRequest] = handleChange.mock.calls[0];
+
+    expect(newParcel.value).toBe(456);
+    expect(changeRequest.nextData.value).toBe(456);
+    expect(changeRequest.prevData.value).toBe(123);
+
+    newParcel.onChange(789);
+
+    let [newParcel2, changeRequest2] = handleChange.mock.calls[1];
+
+    expect(newParcel2.value).toBe(789);
+    expect(changeRequest2.nextData.value).toBe(789);
+    expect(changeRequest2.prevData.value).toBe(456);
+});
+
 test('Parcel.set() should call the Parcels handleChange function with the new parcelData', () => {
     expect.assertions(3);
